requirements.txt
Flask==3.0.0
pymongo==4.6.0
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
pandas==2.1.4
reportlab==4.0.7

